Propagate `THEROCK_BUILD_TESTING` to subprojects
To use `THEROCK_BUILD_TESTING`, the variable must be propagated to the
subprojects. Otherwise, the conditional checking if the RPATH on test
executables should be set silently defaults to false and and the
required shared libraries are not found. This injects
`THEROCK_BUILD_TESTING` by default into the subprojects `_init.cmake`.

Fixes #289