Drop rocprofiler-register on Windows. (#316)
Maintaining this patch is getting harder as more and more Linux-specific
code is getting added, so instead drop the subproject entirely on
Windows instead of trying to gracefully make it no-op. This will push
the responsibility for checking for this dep onto other subprojects
throughout the build system.

Related to https://github.com/ROCm/TheRock/issues/36. Closes
https://github.com/ROCm/TheRock/issues/301.