Create index files and logs for each GPU family
Till now, artifacts got pushed to `therock-artifacts/${{github.run_id}}`
and logs got pushed to `therock-artifacts/${{github.run_id}}/logs`. If
building for multiple GPU-families this resulted in pushing over
pre-existing files as the `github.run_id` is the same for the jobs that
build for different GPU-families. To work around this:

* Create and push an index file for each GPU-family
* Add the GPU-family in the path to the logs

With this the generic artifacts still get replaced by the latest
successful job! This reduces the pressure on storage but logs are
therefore misleading. This should be fine util pushing to the bucket is
moved into scripts.