Refactor workflow to release nightly releases
Refactors the `Portable Linux Package` worflow to build nightly releases
whose assets get continously pushed to a release tagged
`nightly-release`. Furhter allows to push development release to a
release tagged `dev-release`. The latter needs some improvements such as
a build number or maybe the build date to sort the files not only based
on the git hash (which could be shortened).

Progress on #222