[windows] Generate versioned amd_comgr.dll.
I'm not exactly sure how this was ever intended to work and I assume that the old Windows build system was carrying some kind of patch that resolved the situation, because as far as I can tell, the LLVM comgr CMake files have never been retrofitted to suffix the amd_comgr.dll. They have some old code to use the comgr interface version, but we need to use the HIP version, which has to line up from outside of the compiler build.

In any case, making sure these action at a distance things line up seems to be the job of the super-project, so we add code to parse the HIP VERSION (adapted from the clr build), then pass it to the comgr build explicit, and then use a hook to set the name properly. Needless to say, such an arrangement is fairly fragile but seems to be what we are supposed to do.

In the future, I wish we did not have such hard-coded, cross project assumptions and instead just had an explicit config setting to be passed to both to line it up.

Fixes #545