Stop uploading tarballs to GH releases
Tarballs for gfx942 are above the size limit for assets that can be
uploaded to a GH release. With this, the release workflow stops trying
to uploading the tarball, which is also uploaded to S3. This will need
further work as some scripts still try to fetch tarballs from GH.