Append ROCm version to trition version suffix
The `pytorch-triton-rocm` package is build now on a nightly cadence as
will get overwritten as long as the package version is not updated.
While upstream includes a git hash in the version extra, this is not the
case for the ROCm/trition version currently tracked and build alongside
to PyTorch 2.7 from ROCm/pytorch.