Add options to build rccl, rccl-tests and CK from external sources (#1220)
This adds `THEROCK_USE_EXTERNAL_RCCL` and `THEROCK_USE_EXTERNAL_CK`
options to enable building from external sources, not checked out via
TheRock. Enables to use TheRock in CI setups, e.g. to gate PRs, where a
pinned submodule does not fulfill the needs.

Checking if the source directories are set is a little fragile and
checks/erroring is only reliable on fresh configurations but not when
re-configuring an already existing configuration.