[torch][windows] Rebase aotriton patches, fix undeclared `mha_fwd_aot`.
Builds without `--enable-pytorch-flash-attention-windows` were failing with `error: use of undeclared identifier 'mha_fwd_aot'`, but we also needed to remove the `exclude(ATen_HIP_SRCS` code that was added upstream to resolve `lld-link: error: undefined symbol ... transform_bias_rescale_qkv_cuda`.

This is a minimal fix-forward for the first issue and restructures the patches for the second issue.