Use --pre as default instead of specifying version
Add --pre to the default install commands and remove explicit version references. This will expose the prerelase wheels and let pip select the latest version on Linux rather than preferring the non-prerelease wheel. Guidance for installing alternative versions has been modified to communicate that this is Linux-specific and that the dates of the builds should also match.