Exclude `libhipsolver_fortran` from lib
Previously, `libhipsolver_fortran.so` was explicitly excluded from the
lib artifact but included in the test artifact as the shared object is
only required by the `hipsolver-bench` and `hipsolver-test` executables
which are part of the test artifact.

With commit ROCm/hipSOLVER@ddb8256, upstream introduced a soft link to
the library target, resulting in the include and excludes not beeing
applied anymore. This patches should restore the prior partitioning,
avoiding to add an uncessary dependency on Fortran in the lib artifacts
and later the `rocm[libraries]` package.