Reworks pkg-config handling so that it is based on CMAKE_PREFIX_PATH.
Previously, we were manually setting the PKG_CONFIG_PATH environment variable at configure time, but this cannot be reliably propagated to sub-configures that are invoked at build time (as is done for the LLVM runtimes).

This patch switches to adding these prefixes to the CMAKE_PREFIX_PATH, which is the standard way of chaining between top-level and nested configure invocations. It requires a bit of path munging since the CMake pkg-config macros expect to get a prefix, not the fully nested path to the files (which is what the env var expects).

Also inlines a patch to LLVM which fixes an escaping bug for tunneling CMAKE_PREFIX_PATH when it contains more than one entry. I will upstream this to LLVM next week.

Fixes #1481