Restructure CI (build) scripts to harmonize between Win and Linux
The CI to build linux packages has now the same order of execution
as the one for Windows. As part of works for Issue #1456,
"Runner Health Settings" is renamed to "Runner Health Status", installed
python packages are now printed, the rest of the output has been
extended so that Windows and Linux match each other (where applicable).
Commands of this section that modified behavior are moved to different
steps. CCache on Linux has a new section and git commands are moved to
"Fetch Sources". These modifications are also applied to the CI script
for releasing windows packages.