Runs build_tools/.../tests first and removes them from the ctest setup.
It is critical that the build_tools be functional before attempting the build. The way these were, they would be tested after the build succeeded. This adds a lot of cycle time to iteration on build_tools and can make it very hard to track down issues due to broken tools.

This adds on the order of <0.5s to the top of the critical path.

Fixes #1499