[amd_comgr] Implements a delay-loaded trampoline for amd_comgr.
This Linux only addition allows the project to be configured so that amd_comgr is delay loaded into its own dedicated linker namespace installation-wide. This is similar to the previous logic that was only in CLR but extends to any other clients of amd_comgr. It diverges from the existing delay load logic by using the Linux specific `dlmopen` function to manage a standalone namespace for comgr and its deps. This configuration provides the strongest possible isolation of the referenced LLVM deps, eliminating all possible conflicts with other LLVM instances that may be loaded into the process.

* Pulls the Implib.so project into build_tools/third_party for stub generation.
* Adds an amd-comgr-stub project adjacent to amd-comgr.
* [TODO] Add an option to enable delay-loading of amd-comgr. In this mode, we will rename TheRock's amd-comgr project to amd-comgr-impl and amd-comgr-stub to amd-comgr. Then we will only advertise the lib/cmake/amd_comgr_stub packages instead of the default lib/cmake/amd_comgr. In this way, all in-tree projects that take a dep on amd_comgr will get the delay-loading stub transparently.
* [TODO] Patch the clr CMake to enable comgr static linking if the amd_comgr library is STATIC (versus the current logic which switched on BUILD_SHARED_LIBS). This will make CLR operate as if comgr was statically linked and it will no longer be required for maintaining its own dynamic loading trampoline.