Remove 0004-Fix-assembler-error-in-pal-trap-handler.patch
The build system carries a patch to the CRL trap handler for gfx11.
This was to work-around a bug that happened when upgrading the
trap handler from the ROCr: filtering gfx9 support caused an label to be
missing.

The CRL fix for this made sure the required label was defined, while the
patch present in TheRock updated the branch instruction to point to
another label name (placed at the same point in the program).  Those two
fixes are equivalent, so it is un-necessary to continue to carry this
patch in TheRock.