[hipblaslt, rocwmma] Make the empty target default project configurable.
Neither hipblaslt nor rocwmma can build with an explicitly empty target list (the former because Tensile has issues and the latter because hipcc will silently default to gfx906, which is not supported by rocwmma).

There was an existing hack buried in the code to work around hipblaslt. Turns this into a real kwarg to the project declaration and defines it for both.