[CI] Fix SignatureDoesNotMatch by aws credential option (on Windows builds) (#2147)
## Motivation

Fixes #875 which is the issue where Windows builds would fail randomly
when uploading to s3 with the `SignatureDoesNotMatch` error as a result
of special characters existing in the AWS Access Keys generated by the
`configure-aws-credentials` action that is passed through Windows
environment variables to `aws-cli`. More details below.

## Technical Details

https://github.com/ROCm/TheRock/issues/875#issuecomment-3530851762
In summary, in Windows workflows, the `special-characters-workaround`
option is set to true for the `configure-aws-credentials` action which
will regenerate access keys until there are no special characters that
may not be passable through windows environment variables correctly.

## Test Plan

Observe CI.

## Test Result

TBD.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.