Allow users to specify Flang concurrency
On systems with a large number of cores, building Flang can easily use
up all the available memory and swap. To prevent this, allow users to
specify the level of concurrency when building Flang.

This can be done with LLVM's FLANG_PARALLEL_COMPILE_JOBS cmake
variable. We just forward this setting to the amd-llvm build.

If it is not set, the level of concurrency is unlimited (with all of
its consequences).