Python script to promote a version from prerelease to release (#1973)
Allows to promote .tar.gz and wheels from release candidate to release
(e.g. 7.10.0rc1 --> 7.10.0). Can work on a single file, partial matches or
entire directory.

Creates new files for the promotion. Has an option to delete the 
old prerelease version on successful promotion.

As we have a few custom files that also contain the version, added a
modified third-party `change_wheel_version.py` to be able to also 
change the version in those non-standard files inside the wheel.

Works on Windows and Linux. Includes a promotion test.

Issue #1891

---------

Co-authored-by: Scott Todd <scott.todd0@gmail.com>