Add MPFR to common sysdeps
MPFR is a required dependency for GMP and ROCgdb, so include it along with
the other sysdeps.

The sources are cached from the official website. We use the 4.2.2
release.

We do some light smoke testing to make sure the archive and libraries
get built.

MPFR is OS-agnostic, so is kept under third-party/sysdeps/common.

In the future it will be used for Windows-based builds of the debug-tools.