release_portable_linux_pytorch_wheels: set run-name (#2342)
## Motivation

The actions page doesn't make it obvious which job was running which
target. To quickly find the last run of a given target,
extend the run name with that.
<img width="740" height="874" alt="image"
src="https://github.com/user-attachments/assets/7733154d-9d77-49dd-a280-849fdeec76ec"
/>


## Technical Details

Include important job inputs in the name shown on the Actions/workflow
page. This makes it easier to find the last job for a certain target.

## Test Plan

I have used the same trick on other repos; beyond that, github actions
are hard to test.

## Test Result

-

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.