Implement fingerprinting logic required to make artifacts cacheable.
I've been meaning to do this for a year: compute sha256 fingerprints for each subproject and artifact that track the state of all sources, patches, and configuration state.

Adds:

* THEROCK_FPRINT to each subproject target. This is recursively computed from the source state (either git, fetched from external, or patched) in a stable way. It also includes a hash over all configuration settings of the sub-project and its deps.
* Writes an {artifact_name}.fprint file at configure time for every declared artifact. The artifact fprint is computed by combining the fprints for all dependent sub-projects and the descriptor.
* Adds some subproject knobs for special cases:
  * FPRINT_SOURCE_DIR: Overrides the source directory for the purpose of fingerprinting (useful for external fetch).
  * FPRINT_FILE_GLOBS: Additional file globs which should be included in the fingerprint.
  * FPRINT_SOURCE_HASH: Keyword indicating that the source directory should have its contents hashed vs relying on source control state (useful for in-tree, small sub-projects that don't live in an independent git directory.
* Verbose logging when THEROCK_VERBOSE is enabled for debugging fingerprinting issues.
* Logic to the external source fetches (for third-party deps) to write the same metadata as is used for patch management to record stable source hashes (based on URL/hash).

This will be used in CI in combination with a special bootstrap script for certain stages to cache entire artifacts in the remote cache for presubmits. This will be enabled first for base artifacts like the compiler and certain base libraries that change infrequently but are expensive to build and should effectively reduce them to no-ops. We'll evaluate which stages this is suitable for carefully: the further we go in the system the more variable the build system and therefore the higher the risk of accidental cache poisoning, so the first target here is the compiler, which is quite self contained and well managed.