add pytorch smoke test launcher to select GPUs
add pytorch_smoke_test_launcher.py to launch
the pytorch smoke tests.

This launcher will first checks which gpus are
installed on the computer for testing and then do an another
query to check which gpus are supported by the currently running
pytorch version. Then it will combine these lists to get
the index array of GPUs that can be used for running the smoke-tests.

Fixex: https://github.com/ROCm/TheRock/issues/2325

Signed-off-by: Mika Laitio <mika.laitio@amd.com>