Workaround timestamping issues with cmake copies
After downloading and extracting the sources we duplicate the source directory
so we can potentially patch the sources according to our needs.

This cmake-based copy ends up resetting the timestamp of *every* file in
the source directory.

Though this is somewhat harmless, autotools-based projects use timestamps to
recognize when files need to be regenerated. Since we use a release tarball
for GMP we don't need to regenerate anything.

The workaround for this, until we have a better solution, is to touch some
of the documentation files so the timestamp issues don't trigger a call
to makeinfo (unavailable on some systems by default) to regenerate the docs.